Updated use-case factories to combine translation and summarisation use cases. Still have SummarisationUseCaseFactory commented out for testing
