+++ src/app/Main.java
@@ -8,6 +8,7 @@ import interface_adapter.HomeState;
 import interface_adapter.HomeViewModel;
 import interface_adapter.ViewManagerModel;
 import use_case.article_retrieval.ArticleRetrievalDataAccessInterface;
+import use_case.summarization.SummarizationDataAccessInterface;
 import use_case.translation.TranslateAPIDataAccessInterface;
 import data_access.SummarisationDataAccessObject;
 import interface_adapter.*;
@@ -47,18 +48,25 @@ public class Main {
         ArticleViewModel articleViewModel = new ArticleViewModel(new ArticleState());
 
         APIDataAccessObject articleRetrievalDataAccessObject = new APIDataAccessObject();
+        SummarizationDataAccessInterface summarisationDataAccessObject = new SummarisationDataAccessObject();
 
         // Set initial view.
         HomeView homeView = ArticleRetrievalUseCaseFactory.create(viewManagerModel, homeViewModel, articleRetrievalDataAccessObject);
         views.add(homeView, homeView.viewName);
 
-        ArticleView articleView = TranslationUseCaseFactory.create(viewManagerModel, articleViewModel, articleRetrievalDataAccessObject);
+        ArticleView articleView = TranslationUseCaseFactory.create(
+                viewManagerModel,
+                articleViewModel,
+                articleRetrievalDataAccessObject,
+                summarisationDataAccessObject
+        );
+
         views.add(articleView, articleView.viewName);
 
 //        // TODO: FOR ARTICLEVIEW TESTING PURPOSES ONLY, COMMENT THIS PART OUT WHEN DONE!
-//        viewManagerModel.setActiveView(articleView.viewName);
+        viewManagerModel.setActiveView(articleView.viewName);
 
-        viewManagerModel.setActiveView(homeView.viewName);
+//        viewManagerModel.setActiveView(homeView.viewName);
         viewManagerModel.firePropertyChanged();
 
         application.pack();
