+++ src/app/TranslationUseCaseFactory.java
@@ -2,9 +2,15 @@ package app;
 
 import interface_adapter.ArticleViewModel;
 import interface_adapter.ViewManagerModel;
+import interface_adapter.summarization.SummarizationController;
+import interface_adapter.summarization.SummarizationPresenter;
 import interface_adapter.translation.TranslationController;
 import interface_adapter.translation.TranslationPresenter;
 import interface_adapter.translation.TranslationPresenter;
+import use_case.summarization.SummarizationDataAccessInterface;
+import use_case.summarization.SummarizationInputBoundary;
+import use_case.summarization.SummarizationInteractor;
+import use_case.summarization.SummarizationOutputBoundary;
 import use_case.translation.TranslateAPIDataAccessInterface;
 import use_case.translation.TranslationInputBoundary;
 import use_case.translation.TranslationInteractor;
@@ -13,9 +19,13 @@ import view.ArticleView;
 
 public class TranslationUseCaseFactory {
 
-    public static ArticleView create(ViewManagerModel viewManagerModel, ArticleViewModel articleViewModel, TranslateAPIDataAccessInterface translationDataAccessObject) {
+    public static ArticleView create(ViewManagerModel viewManagerModel,
+                                     ArticleViewModel articleViewModel,
+                                     TranslateAPIDataAccessInterface translationDataAccessObject,
+                                     SummarizationDataAccessInterface summarizationDataAccessObject) {
         TranslationController translationController = createTranslationUseCase(viewManagerModel, articleViewModel, translationDataAccessObject);
-        return new ArticleView(translationController, articleViewModel);
+        SummarizationController summarizationController = createSummarisationUseCase(viewManagerModel, articleViewModel, summarizationDataAccessObject);
+        return new ArticleView(translationController, articleViewModel, summarizationController);
     }
 
     private static TranslationController createTranslationUseCase(ViewManagerModel viewManagerModel, ArticleViewModel articleViewModel, TranslateAPIDataAccessInterface translationDataAccessObject) {
@@ -23,5 +33,12 @@ public class TranslationUseCaseFactory {
         TranslationInputBoundary translationInputBoundary = new TranslationInteractor(translationOutputBoundary, translationDataAccessObject);
 
         return new TranslationController(translationInputBoundary);
+    }
+
+    private static SummarizationController createSummarisationUseCase(ViewManagerModel viewManagerModel, ArticleViewModel articleViewModel, SummarizationDataAccessInterface summarizationDataAccessObject) {
+        SummarizationOutputBoundary summarisationController = new SummarizationPresenter(articleViewModel);
+        SummarizationInputBoundary summarizationInteractor = new SummarizationInteractor(summarisationController, summarizationDataAccessObject);
+
+        return new SummarizationController(summarizationInteractor);
     }
 }
